requirements.txt
PyQt6>=6.5
Pillow>=10.0
pyinstaller>=6.3

